delete tax and discount row. and refactoring the code of add/delete tax_and_discount fields  and calculate subtotal actions of invoice controller.
