+++ app/controllers/invoices_controller.rb
@@ -17,48 +17,6 @@ class InvoicesController < ApplicationController
     @invoice = Invoice.new(user: current_user, status: Invoice::DRAFT)
   end
 
-  def add_td_fields
-    @tax_and_discount = TaxAndDiscount.find(params[:taxAndDiscountId])
-    @tax_and_discount_poly = TaxAndDiscountPoly.new(
-      name: @tax_and_discount.name,
-      amount: @tax_and_discount.amount,
-      td_type: @tax_and_discount.td_type,
-      tax_and_discount_id: @tax_and_discount.id,
-      tax_type: @tax_and_discount.tax_type
-    )
-  end
-
-  def delete_td_fields
-    @tax_and_discount_poly = TaxAndDiscountPoly.find_by(id: params[:taxAndDiscountId])
-    @tax_and_discount_poly.destroy if @tax_and_discount_poly.present? && @tax_and_discount_poly.persisted?
-  end
-
-  # this action is used at three scenarios
-  # 1. when user adds/removes a line item
-  # 2. when user updates the quantity of line item
-  # 3. when user updates the unit rate of line item
-  def calculate_sub_total
-    line_items = params[:lineItemsAttributes].map do |line_item|
-      if line_item[:quantity].present? && line_item[:unitRate].present?
-        LineItem.new(
-          quantity: line_item[:quantity].to_f,
-          unit_rate: line_item[:unitRate].to_f
-        )
-      end
-    end
-
-    tax_and_discount_polys = params[:taxAndDiscountPolyAttributes].map do |td_poly|
-      TaxAndDiscountPoly.new(
-        amount: td_poly[:amount].to_f,
-        td_type: td_poly[:td_type],
-        name: td_poly[:name],
-        tax_type: td_poly[:tax_type]
-      )
-    end
-    
-    @sub_total = ::InvoiceAmountCalculator.new.calculate_sub_total(line_items, tax_and_discount_polys)
-  end
-
   # GET /invoices/1/edit
   def edit
   end
