+++ app/models/concerns/tax_and_discount_validations.rb
@@ -1,5 +1,6 @@
 module TaxAndDiscountValidations
   extend ActiveSupport::Concern
+  
   FIXED = 'fixed'
   PERCENTAGE = 'percentage'
   TAX = 'tax'
@@ -10,11 +11,11 @@ module TaxAndDiscountValidations
     enum td_type: {tax: TAX, discount: DISCOUNT}
   end
 
-  def tax?
+  def tax
     amount > 0
   end
 
-  def discount?
+  def discount
     amount < 0
   end
 end
