+++ app/services/invoice_amount_calculator.rb
@@ -4,18 +4,18 @@ class InvoiceAmountCalculator
     
     tax_total = 0
     discount_total = 0
+    
     tax_and_discount_polys.each do |td_poly|
       if td_poly.tax_type == 'percentage'
-        if td_poly.tax?
+        if td_poly.tax
           tax_total += line_items_total * (td_poly.amount / 100.0)
-        elsif td_poly.discount?
+        elsif td_poly.discount
           discount_total += line_items_total * (td_poly.amount / 100.0)
         end
-
       elsif td_poly.tax_type == 'fixed'
-        if td_poly.tax?
+        if td_poly.tax
           tax_total += td_poly.amount
-        elsif td_poly.discount?
+        elsif td_poly.discount
           discount_total += td_poly.amount
         end
       end
