--- app/views/invoices/add_td_fields.turbo_stream.erb
@@ -1,4 +1,0 @@
-<%= turbo_stream.append "tax_and_discount_poly" do %>
-  <%= render partial: "td_poly_fields", locals: { tax_and_discount_poly: @tax_and_discount_poly } %>
-<% end %>
-
