+++ app/views/invoices/tax_and_discount_polies/create.turbo_stream.erb
@@ -1,0 +1,3 @@
+<%= turbo_stream.append "tax_and_discount_poly" do %>
+  <%= render partial: "invoices/td_poly_fields", locals: { tax_and_discount_poly: @tax_and_discount_poly } %>
+<% end %>
